requirements.txt
PyPDF2
python-docx
docx2txt
python-pptx
